Remove ruby 2.0.0 strict version dependency
According to the bundler docs - http://bundler.io/v1.5/gemfile_ruby.html - the ruby version directive is for strict binary compatibility, and is therefore too strict for our purposes. In fact we need any version of ruby which can run the github-pages gem, which is already enforced by the following line.