Replace second front page example
This commit replaces the second front page example with a new original
one, titled "60 Hz Gabber Rave 1995." The previous example had to go
since it used the Atari2600 UGen, which is not in either core or
sc3-plugins.

The example is my original work. I release it under the public domain.